+++ test/mock.js
@@ -14,7 +14,9 @@ var Mock = {
             isUserLogged: function(){ Mock.calls.push('isUserLogged'); return Mock.logged; },
             rankContent: function(){ Mock.calls.push('rankContent'); },
             getUserToken: function(){ Mock.calls.push('getUserToken'); },
-            setUserStateChangeListener: function(callback){ Mock.calls.push('setUserStateChangeListener'); callback.call(); },
+            setUserStateChangeListener: function(callbacks){
+                Mock.calls.push('setUserStateChangeListener'); callbacks.onLoginStateChange(); 
+            },
             userLogout: function(){ Mock.calls.push('userLogout'); },
             finalizeLoginFlow: function(callback){ Mock.calls.push('finalizeLoginFlow'); callback.call(); },
             // login methods
@@ -55,7 +57,19 @@ var Mock = {
             setOnForgotFlowCallback: function(callback){ Mock.calls.push('setOnForgotFlowCallback'); callback.call(); return this; },
             getMSISDNPINForgotFlow: function(){ Mock.calls.push('getMSISDNPINForgotFlow'); return this; },
             startForgotFlow: function(){ Mock.calls.push('startForgotFlow'); return this; },
-            setLogViewInfo: function(){ Mock.calls.push('setLogViewInfo'); return this; }
+            setLogViewInfo: function(){ Mock.calls.push('setLogViewInfo'); return this; },
+            setUserProperties: function () { Mock.calls.push('setUserProperties'); return this; },
+            getEmailSignupFlow: function () { Mock.calls.push('getEmailSignupFlow'); return this; },
+            getPaymentManager: function() { 
+                return {
+                    getOfferFor: function(str, str2, fn) {
+                        fn(null, 'offerIdMock');
+                    },
+                    addSerializedPayment: function(str, fn) {
+                        fn();
+                    }
+                };
+            }
         };
         Mock.Newton = {
             getSharedInstanceWithConfig: function(){ Mock.calls.push('getCustomFlow'); return Mock.NewtonMock; },
@@ -115,6 +129,8 @@ var Mock = {
         spyOn(Mock.NewtonMock, 'getMSISDNPINForgotFlow').and.callThrough();
         spyOn(Mock.NewtonMock, 'startForgotFlow').and.callThrough();
         spyOn(Mock.NewtonMock, 'setLogViewInfo').and.callThrough();
+        spyOn(Mock.NewtonMock, 'setUserProperties').and.callThrough();
+        spyOn(Mock.NewtonMock, 'getEmailSignupFlow').and.callThrough();
         
         spyOn(Mock.Newton, 'getSharedInstanceWithConfig').and.callThrough();
     }